+++ test/compute-engine/compile.test.ts
@@ -1,0 +1,78 @@
+import { engine as ce } from '../utils';
+
+describe('COMPILE', () => {
+  describe('Expressions', () => {
+    it('should compile (and simplify) a simple expression', () => {
+      expect(
+        ce.parse('3.45 + \\frac57').compile()?.toString()
+      ).toMatchInlineSnapshot(`0.7142857142857143 + 3.45`);
+    });
+
+    it('should compile an expression with a constant', () => {
+      expect(
+        ce.parse('2\\exponentialE').compile()?.toString()
+      ).toMatchInlineSnapshot(`2 * Math.E`);
+    });
+
+    it('should compile an expression with trig functions', () => {
+      expect(
+        ce.parse('2 \\cos(\\frac{\\pi}{5})').compile()?.toString()
+      ).toMatchInlineSnapshot(`2 * (Math.sqrt(5) / 4 + 0.25)`);
+    });
+  });
+
+  describe('Blocks', () => {
+    it('should compile a simple block', () => {
+      const expr = ce.box(['Block', ['Multiply', 10, 2]]);
+      expect(expr.compile()?.toString() ?? '').toMatchInlineSnapshot(`
+        (() => {
+        return 10 * 2
+        })()
+      `);
+    });
+
+    it('should compile a block with two statements', () => {
+      const expr = ce.box(['Block', ['Add', 13, 15], ['Multiply', 10, 2]]);
+      expect(expr.compile()?.toString() ?? '').toMatchInlineSnapshot(`
+        (() => {
+        13 + 15;
+        return 10 * 2
+        })()
+      `);
+    });
+
+    it('should compile a block with a declaration', () => {
+      const expr = ce.box([
+        'Block',
+        ['Declare', 'x', 'Numbers'],
+        ['Assign', 'x', 4.1],
+        ['Multiply', 'x', 'n'],
+      ]);
+      expect(expr.compile()?.toString() ?? '').toMatchInlineSnapshot(`
+        (() => {
+        let x;
+        x = 4.1;
+        return x * _.n
+        })()
+      `);
+    });
+
+    it('should compile a block with a return statement', () => {
+      const expr = ce.box([
+        'Block',
+        ['Declare', 'x', 'Numbers'],
+        ['Assign', 'x', 4.1],
+        ['Return', ['Add', 'x', 1]],
+        ['Multiply', 'x', 2],
+      ]);
+      expect(expr.compile()?.toString() ?? '').toMatchInlineSnapshot(`
+        (() => {
+        let x;
+        x = 4.1;
+        return x + 1;
+        return x * 2
+        })()
+      `);
+    });
+  });
+});